# docs/swagger.yaml
basePath: /
definitions:
...
  dto.BaseResponse:
    properties:
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.DepartmentResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      message:
        type: string
      name:
        type: string
      success:
        example: true
        type: boolean
      updated_at:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      success:
        example: false
        type: boolean
    type: object
  dto.HospitalAdminChangeRoleRequest:
    properties:
      role:
        allOf:
        - $ref: '#/definitions/entity.UserRole'
        example: RECEIVING_SPECIALIST
    required:
    - role
    type: object
  dto.HospitalAdminCreateStaffRequest:
    properties:
      department_id:
        example: dfc2b777-a5d5-424b-911a-976b2e8d8614
        type: string
      email:
        example: staff@hospital.et
        type: string
      first_name:
        example: Abebe
        type: string
      last_name:
        example: Molla
        type: string
      middle_name:
        example: Kebede
        type: string
      national_id:
        example: ETH-0001
        type: string
      password:
        example: password123
        minLength: 8
        type: string
      role:
        allOf:
        - $ref: '#/definitions/entity.UserRole'
        example: LIAISON_OFFICER
    required:
    - email
    - first_name
    - last_name
    - middle_name
    - password
    - role
    type: object
  dto.HospitalAdminReplaceStaffRequest:
    properties:
      email:
        example: replacement@hospital.et
        type: string
      first_name:
        example: Abebe
        type: string
      last_name:
        example: Molla
        type: string
      middle_name:
        example: Kebede
        type: string
      password:
        example: newStrongPass123
        minLength: 8
        type: string
      reason:
        example: Staff transition and account continuity
        minLength: 5
        type: string
    required:
    - email
    - first_name
    - last_name
    - middle_name
    - password
    - reason
    type: object
...
  dto.HospitalResponse:
    properties:
      address:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      message:
        type: string
      name:
        type: string
      region:
        type: string
      success:
        example: true
        type: boolean
      tier_level:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.LogResponseDTO:
    properties:
      changed_by_id:
        type: string
      created_at:
        type: string
      from_status:
        type: string
      history_id:
        type: string
      referral_id:
        type: string
      role:
        type: string
      to_status:
        type: string
    type: object
...
  dto.PaginatedLogResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.LogResponseDTO'
        type: array
      message:
        type: string
      page:
        type: integer
      page_size:
        type: integer
      success:
        example: true
        type: boolean
      total:
        type: integer
    type: object
...
  dto.UserListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      message:
        type: string
      page:
        type: integer
      success:
        example: true
        type: boolean
      total:
        type: integer
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      department:
        $ref: '#/definitions/dto.DepartmentResponse'
      department_id:
        type: string
      email:
        type: string
      first_name:
        type: string
      hospital:
        $ref: '#/definitions/dto.HospitalResponse'
      hospital_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      message:
        type: string
      middle_name:
        type: string
      national_id:
        type: string
      profile_image_url:
        type: string
      role:
        $ref: '#/definitions/entity.UserRole'
      success:
        example: true
        type: boolean
      updated_at:
        type: string
    type: object
...
  entity.UserRole:
    enum:
    - REFERRING_DOCTOR
    - LIAISON_OFFICER
    - RECEIVING_SPECIALIST
    - RECEPTIONIST
    - MOH_ANALYST
    - DEPT_HEAD
    - HOSPITAL_ADMIN
    - SYSTEM_SUPER_ADMIN
    type: string
    x-enum-varnames:
    - RoleReferringDoctor
    - RoleLiaisonOfficer
    - RoleReceivingSpecialist
    - RoleReceptionist
    - RoleMohAnalyst
    - RoleDeptHead
    - RoleHospitalAdmin
    - RoleSystemSuperAdmin
...
info:
  contact: {}
  description: |-
    # Hospital Referral Hub API
    A national‑scale hospital referral management platform that digitises the entire patient‑transfer workflow, from initial doctor referral to final clinical outcome. All actions are governed by strict role‑based access controls and clinical governance rules.

    ---
    ## Referral Lifecycle
    DRAFT → SUBMITTED → (Liaison) UNDER_LIAISON_REVIEW → FORWARDED
    ↘ REJECTED_BY_LIAISON
    FORWARDED → (Specialist) UNDER_SPECIALIST_REVIEW → ACCEPTED / REJECTED_BY_SPECIALIST
    ACCEPTED → SCHEDULED → ASSIGNED → COMPLETED
    ↘ MISSED / RESCHEDULED / DECEASED

    ---
    ## Visibility Rules

    | Status | Visible To |
    |----------------------|-----------|
    | DRAFT / NEED_REVISION | Referring Doctor only |
    | SUBMITTED … FORWARDED | Liaison of the sender hospital |
    | FORWARDED … COMPLETED | Specialists of the target hospital |
    | ACCEPTED … RESCHEDULED| Receptionists of the target hospital |
    | All statuses | System Admins (global); MoH Analysts (aggregated dashboards, no raw clinical data) |

    ---
    ## Critical Business Rules

    - **ML Triage Gate**: A referral cannot be accepted without a severity score (set manually via `POST /specialist/referrals/{id}/triage-severity`).
...
    - **Walk‑in Restriction**: Walk‑ins can only be registered for referrals in ACCEPTED or SCHEDULED status.
...

    For detailed per‑endpoint rules, see the individual endpoint descriptions below.
...
paths:
...
  /api/v1/hospital-admin/referrals/{id}/status-history:
    get:
      description: |-
        Read-only status history for a referral linked to the admin's hospital.
        **Roles:** HOSPITAL_ADMIN
        **Prerequisites:** Referral must be linked to the admin's hospital.
        **Common Errors:**
        - 401 Unauthorized
        - 403 (out of scope)
      parameters:
      - description: Referral ID
        in: path
        name: id
        required: true
        type: string
      - default: 20
        description: Pagination limit
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedLogResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get referral status history (Hospital Admin)
      tags:
      - Hospital Admin - Staff Management
  /api/v1/hospital-admin/staff:
    get:
      description: |-
        List staff users scoped to the admin's hospital.
        **Roles:** HOSPITAL_ADMIN
        **Prerequisites:** Admin must belong to a hospital.
        **Common Errors:**
        - 401 Unauthorized
        - 403 Forbidden
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by role
        in: query
        name: role
        type: string
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by email
        in: query
        name: email
        type: string
      - description: Filter by department ID
        in: query
        name: dept_id
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserListResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List staff (Hospital Admin)
      tags:
      - Hospital Admin - Staff Management
    post:
      consumes:
      - application/json
      description: |-
        Create a new staff user in the current hospital scope.
        **Roles:** HOSPITAL_ADMIN
        **Prerequisites:** Admin must belong to a hospital.
        **Common Errors:**
        - 400 invalid input
        - 403 (wrong hospital scope)
        - 409 (duplicate email/national_id)
      parameters:
      - description: Staff creation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.HospitalAdminCreateStaffRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create staff (Hospital Admin)
      tags:
      - Hospital Admin - Staff Management
  /api/v1/hospital-admin/staff/{id}:
    delete:
      description: |-
        Soft delete a staff user within the same hospital scope.
        **Roles:** HOSPITAL_ADMIN
        **Prerequisites:** Target user must belong to the same hospital.
        **Common Errors:**
        - 403 (cannot delete self or out of scope)
        - 404 Not Found
      parameters:
      - description: Staff user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Soft delete staff (Hospital Admin)
      tags:
      - Hospital Admin - Staff Management
    get:
      description: |-
        Get a staff profile scoped to the admin's hospital.
        **Roles:** HOSPITAL_ADMIN
        **Prerequisites:** Target user must belong to the same hospital.
        **Common Errors:**
        - 400 Invalid ID format
        - 403 (out of scope)
        - 404 Not Found
      parameters:
      - description: Staff user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get staff by ID (Hospital Admin)
      tags:
      - Hospital Admin - Staff Management
  /api/v1/hospital-admin/staff/{id}/replace:
    post:
      consumes:
      - application/json
      description: |-
        In-place replacement updates identity and password while keeping the same user ID.
        **Roles:** HOSPITAL_ADMIN
        **Prerequisites:** Target user must belong to the same hospital.
        **Common Errors:**
        - 400 invalid input
        - 403 (out of scope)
        - 409 (conflict)
      parameters:
      - description: Staff user ID
        in: path
        name: id
        required: true
        type: string
      - description: Replacement payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.HospitalAdminReplaceStaffRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Replace staff in-place (Hospital Admin)
      tags:
      - Hospital Admin - Staff Management
  /api/v1/hospital-admin/staff/{id}/role:
    patch:
      consumes:
      - application/json
      description: |-
        Update the role of a staff user within the same hospital scope.
        **Roles:** HOSPITAL_ADMIN
        **Prerequisites:** Target user must belong to the same hospital.
        **Common Errors:**
        - 400 invalid role or input
        - 403 (out of scope)
      parameters:
      - description: Staff user ID
        in: path
        name: id
        required: true
        type: string
      - description: Role change payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.HospitalAdminChangeRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change staff role (Hospital Admin)
      tags:
      - Hospital Admin - Staff Management
...
securityDefinitions:
  BearerAuth:
    description: Enter your bearer token in the format **Bearer &lt;token&gt;**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"